+++ test.py
@@ -74,8 +74,7 @@ def debug_case_9():
 
 def debug_case_5_all():
     points = matcher.load_points('./data/case_5_all.geojson')
-    # tmp = points[50:].reset_index()
-    tmp = points[0:].reset_index()
+    tmp = points[55:].reset_index()
 
     # tmp = matcher.load_points("./data/case_6.geojson")
 
@@ -85,6 +84,74 @@ def debug_case_5_all():
     res = matcher.matching(tmp, simplify=False, plot=True, details=True, dir_trans=False, debug_in_levels=False, top_k=5, metric='lcss')
     print(res['epath'])
 
-debug_case_5_all()
+    return res
+
+def debug_duplicate_two_points():
+    #! traj_3 两个重叠点
+    points = matcher.load_points('./data/trajs/traj_3.geojson')
+    plot_geodata(points)
+    # points = check_duplicate_points(points)
+    tmp = points.reset_index()
+
+    # tmp = matcher.load_points("./data/case_6.geojson")
+
+    revert = False
+    if revert:
+        tmp = tmp[::-1].reset_index(drop=True)
+
+    res = matcher.matching(tmp, simplify=False, plot=True, details=True, dir_trans=False, debug_in_levels=False, top_k=5, metric='lcss')
+    # print(res['epath'])
+
+    return res
+
+# res = debug_case_5_all()
+# res = debug_duplicate_two_points()
 
 #%%
+points = matcher.load_points('./data/trajs/traj_9.geojson')
+plot_geodata(points)
+
+# points = check_duplicate_points(points)
+tmp = points[:50].reset_index()
+
+# tmp = matcher.load_points("./data/case_6.geojson")
+
+revert = False
+if revert:
+    tmp = tmp[::-1].reset_index(drop=True)
+
+res = matcher.matching(tmp, simplify=True, plot=False, details=True, dir_trans=False, 
+                       debug_in_levels=False, top_k=5, metric='lcss')
+
+matcher.plot_result(tmp, res)
+
+# %%
+res['details']['steps']
+
+# %%
+res['details']['graph'].loc[4]
+
+
+res['details']['cands'][['pid', 'eid']].groupby('pid').eid.apply(list).apply(np.sort)
+
+res['details']['graph'].loc[3].index.get_level_values(1).unique()
+
+
+# res['details']['simplified_traj'].PID.values
+
+# %%
+
+f_score = [
+    {105275: 0.9968586964007947, 63894: 0.574929140140689, 105276: 0.5680360182757898, 63895: 0.5680360182757898}, 
+    {95697: 0.9968586964007947, 95698: 0.9968586964007947, 105275: 0.9968586964007947}, 
+    {105275: 1.4788528072730815}, 
+    {95697: 1.4788528072730815, 95698: 1.4788528072730815, 105275: 1.4788528072730815}]
+
+prev_path = {
+    0: {105275: None, 63894: None, 105276: None, 63895: None}, 
+    1: {95697: (0, 105275), 95698: (0, 105275), 105275: (0, 105275)}, 
+    2: {105275: (1, 95697)}, 
+    3: {95697: (2, 105275), 95698: (2, 105275), 105275: (2, 105275)}}
+
+
+# %%